+++ Kyoo.Common/Models/LibraryLink.cs
@@ -2,12 +2,12 @@ namespace Kyoo.Models
 {
 	public class LibraryLink
 	{
-		public long ID { get; set; }
-		public long LibraryID { get; set; }
+		public int ID { get; set; }
+		public int LibraryID { get; set; }
 		public virtual Library Library { get; set; }
-		public long? ShowID { get; set; }
+		public int? ShowID { get; set; }
 		public virtual Show Show { get; set; }
-		public long? CollectionID { get; set; }
+		public int? CollectionID { get; set; }
 		public virtual Collection Collection { get; set; }
 
 		public LibraryLink() { }
