+++ Kyoo.Common/Models/MetadataID.cs
@@ -4,20 +4,20 @@ namespace Kyoo.Models
 {
 	public class MetadataID
 	{
-		[JsonIgnore] public long ID { get; set; }
-		[JsonIgnore] public long ProviderID { get; set; }
+		[JsonIgnore] public int ID { get; set; }
+		[JsonIgnore] public int ProviderID { get; set; }
 		public virtual ProviderID Provider {get; set; }
 		
-		[JsonIgnore] public long? ShowID { get; set; } 
+		[JsonIgnore] public int? ShowID { get; set; } 
 		[JsonIgnore] public virtual Show Show { get; set; }
 		
-		[JsonIgnore] public long? EpisodeID { get; set; } 
+		[JsonIgnore] public int? EpisodeID { get; set; } 
 		[JsonIgnore] public virtual Episode Episode { get; set; }
 		
-		[JsonIgnore] public long? SeasonID { get; set; } 
+		[JsonIgnore] public int? SeasonID { get; set; } 
 		[JsonIgnore] public virtual Season Season { get; set; }
 		
-		[JsonIgnore] public long? PeopleID { get; set; } 
+		[JsonIgnore] public int? PeopleID { get; set; } 
 		[JsonIgnore] public virtual People People { get; set; }
 		
 		public string DataID { get; set; }
